+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+- 0.2.1.0
+    - Fix `Traversable`, `TraversableWithIndex`, `FoldableWithIndex` to traverse
+      in insertion order
+    - Add `unorderedTraverse`, `unorderedTraverseWithKey`, `unoderedFoldMap`, and
+      `unorderedFoldMapWithKey`.
+
 - 0.2.0.0
     - Use `aeson-1`
     - removed our `FromJSONKey` and `ToJSONKey` in favour of `aeson` variants